requirements.txt
# ML/Data Science
torch>=1.9.0  # For deep learning modules
tensorflow
jax
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
scikit-learn>=0.24.0

# Bioinformatics
biopython>=1.79
mdtraj
prody
biotite

# Visualization
plotly
py3dmol

# Web
flask>=2.0.0

# NLP
transformers>=4.5.0  # For LLM integration
datasets
tokenizers

# Simulation
openmm

# Utilities
requests>=2.26.0
pytest>=6.2.5
pytest-cov>=2.12.0
black>=21.7b0  # For code formatting
flake8>=3.9.0  # For linting
mypy>=0.910  # For type checking
